Refactor replay history internals
